% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{create_empty_enrollment_df}
\alias{create_empty_enrollment_df}
\title{Create empty enrollment data frame}
\usage{
create_empty_enrollment_df(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Empty data frame with standard columns
}
\description{
Returns an empty data frame with the standard schema.
}
\keyword{internal}
