% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_available_years}
\alias{detect_available_years}
\title{Detect available years from EdSight}
\usage{
detect_available_years(min_year, max_year)
}
\arguments{
\item{min_year}{Minimum year to check}

\item{max_year}{Maximum year to check}
}
\value{
Integer vector of available years
}
\description{
Queries the EdSight portal to determine which years have data available.
}
\keyword{internal}
