% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_edsight_enrollment}
\alias{download_edsight_enrollment}
\title{Download enrollment data from EdSight}
\usage{
download_edsight_enrollment(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
List with enrollment data frames
}
\description{
Connecticut's EdSight portal provides enrollment data exports.
This function attempts to download data via available API endpoints.
}
\keyword{internal}
