% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{fetch_education_directory}
\alias{fetch_education_directory}
\title{Fetch Connecticut education directory}
\usage{
fetch_education_directory()
}
\value{
Data frame with organization information
}
\description{
Downloads the current list of schools and districts from CT Open Data.
}
\keyword{internal}
