% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_directory_cache_path}
\alias{get_directory_cache_path}
\title{Get directory cache file path}
\usage{
get_directory_cache_path()
}
\value{
Full path to directory cache file
}
\description{
Get directory cache file path
}
\keyword{internal}
