% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_raw_directory}
\alias{get_raw_directory}
\title{Get raw school directory data from CT Open Data}
\usage{
get_raw_directory()
}
\value{
Raw data frame as downloaded from CT Open Data
}
\description{
Downloads the raw school directory data from the Connecticut Open Data Portal
via the Socrata API.
}
\keyword{internal}
