% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_enr}
\alias{process_enr}
\title{Process raw enrollment data}
\usage{
process_enr(raw_data, end_year)
}
\arguments{
\item{raw_data}{List of raw data frames from get_raw_enr()}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Transforms raw enrollment data from CSDE into standardized format.
}
\keyword{internal}
