% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_enrollment.R
\name{process_local_enrollment}
\alias{process_local_enrollment}
\title{Process locally imported enrollment data}
\usage{
process_local_enrollment(df, end_year)
}
\arguments{
\item{df}{Raw data frame from file import}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Handles various formats of EdSight enrollment exports.
}
\keyword{internal}
