% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{read_directory_cache}
\alias{read_directory_cache}
\title{Read directory data from cache}
\usage{
read_directory_cache()
}
\value{
Cached data frame
}
\description{
Read directory data from cache
}
\keyword{internal}
