% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{standardize_org_code}
\alias{standardize_org_code}
\title{Standardize district/school identifiers}
\usage{
standardize_org_code(code)
}
\arguments{
\item{code}{Organization code}
}
\value{
Standardized 7-character code
}
\description{
Connecticut uses 7-digit organization codes.
}
\keyword{internal}
