% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_enrollment.R
\name{tidy_wide_data}
\alias{tidy_wide_data}
\title{Tidy wide-format enrollment data}
\usage{
tidy_wide_data(df)
}
\arguments{
\item{df}{A wide data.frame with demographic columns and grade_XX columns}
}
\value{
A long data.frame of tidied enrollment data
}
\description{
Transforms wide enrollment data to long format with subgroup column.
}
\keyword{internal}
