% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{try_ct_open_data}
\alias{try_ct_open_data}
\title{Try fetching enrollment data from CT Open Data portal}
\usage{
try_ct_open_data(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame or NULL if not found
}
\description{
Checks for enrollment datasets on data.ct.gov via Socrata API.
}
\keyword{internal}
