% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{write_directory_cache}
\alias{write_directory_cache}
\title{Write directory data to cache}
\usage{
write_directory_cache(df)
}
\arguments{
\item{df}{Data frame to cache}
}
\value{
Invisibly returns the cache path
}
\description{
Write directory data to cache
}
\keyword{internal}
