% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_raw_directory}
\alias{get_raw_directory}
\title{Get raw directory data from CT Open Data API}
\usage{
get_raw_directory()
}
\value{
Raw data frame from API response
}
\description{
Downloads the complete Education Directory dataset from data.ct.gov.
}
\keyword{internal}
